Updated the ApplyCOnfig System
The new system use only one single file for all controls.
For each object key;value to attribute, the code check if a set parameter is available at the name of the key, if so then its a valid element parameter and then its saved & sent to the UI to update render.

This probably removed around 500 lines of repetitive code that nobody want, aswell as clearing the code itself.

I also switched some ui elements as update, the UI is now considered as an element with a tag for actions such as x,y, width, lenght that are the same arguments than controllers.